requirements.txt
django==3.2.4
whitenoise==5.2.0
gunicorn==20.0.4
psycopg2==2.8
dj-database-url
python-decouple
pyshp==2.1.3
pygithub==1.55
